Pass the svg file name as an argument
Instead of hard coding it into the script. Standard stuff.

+++ export-layers.sh
@@ -1,12 +1,19 @@
 #!/bin/bash -e
 
-layers=$(xmlstarlet sel -t -m "//_:svg/_:g" -v "@id" -n all-slides.svg)
+inputSvgFile=$1
+
+if [ -z "$inputSvgFile" ]; then
+	echo "Usage: export-layers.sh <svg file>"
+	exit 1
+fi
+
+layers=$(xmlstarlet sel -t -m "//_:svg/_:g" -v "@id" -n $inputSvgFile)
 
 mkdir -p build
 rm -f build/layers
 
 for layer in $layers; do
-  label=$(xmlstarlet sel -t -m "//_:svg/_:g[@id='$layer']" -v "@inkscape:label" -n all-slides.svg)
+  label=$(xmlstarlet sel -t -m "//_:svg/_:g[@id='$layer']" -v "@inkscape:label" -n $inputSvgFile)
   label=$(echo $label | sed 's/ /-/g')
   echo $label >> build/layers
 done
@@ -14,9 +21,9 @@ done
 cat build/layers  | jq  --raw-input .  | jq --slurp . > src/names.json
 
 for layer in $layers; do
-  label=$(xmlstarlet sel -t -m "//_:svg/_:g[@id='$layer']" -v "@inkscape:label" -n all-slides.svg)
+  label=$(xmlstarlet sel -t -m "//_:svg/_:g[@id='$layer']" -v "@inkscape:label" -n $inputSvgFile)
   label=$(echo $label | sed 's/ /-/g')
-  params=$(inkscape --query-id $layer --query-x --query-y --query-width --query-height all-slides.svg)
+  params=$(inkscape --query-id $layer --query-x --query-y --query-width --query-height $inputSvgFile)
 
   x=$(echo $params | cut -d ' ' -f 1)
   y=$(echo $params | cut -d ' ' -f 2)
@@ -25,6 +32,6 @@ for layer in $layers; do
   posX=$(python -c "print int($x + $width/2)")
   posY=$(python -c "print int($y + $height/2)")
   echo ", ( \"$label\", { x = $posX, y = $posY } )"
-  inkscape --export-area-page --export-id="$layer" --export-id-only --export-filename="public/layers/$label.svg" --export-type=svg all-slides.svg
+  inkscape --export-area-page --export-id="$layer" --export-id-only --export-filename="public/layers/$label.svg" --export-type=svg $inputSvgFile
 done
